Improve menu repeat input typing (#13993)
* Improve ArtiCtrlCur pad repeat typing

* Improve CmdCtrlCur pad repeat typing

* Improve menu repeat input typing

---------

+++ src/menu_cmd.cpp
@@ -1288,7 +1288,7 @@ unsigned int CMenuPcs::CmdCtrlCur()
 {
 	bool blocked = false;
 	unsigned int press;
-	unsigned int hold;
+	u16 hold;
 	int caravanWork = Game.m_scriptFoodBase[0];
 
 	if ((Pad._452_4_ != 0) || (Pad._448_4_ != -1)) {
